refactor: extract daemon child bootstrap composition
